Add code for signing AWS requests.  At its core, aws_sign computes AWS4 signatures; wrapped around that are simple routines which sign S3, EC2, SNS, and DynamoDB requests.
Code for signing requests to other AWS services will probably be
added in the future, on an as-needed basis.

This originated in 2014 as code for signing S3 requests in an
unreleased version of kivaloo and has been gradually extended since
then; it was first released as part of bsdec2-image-upload.